+++ plugins/image-trust/pkg/config/modes.go
@@ -1,8 +1,10 @@
 package config
 
 const (
-	ModeCosignKeyless = "cosign-keyless"
-	ModeCosignKey     = "cosign-key"
+	ModeCosignKeyless            = "cosign-keyless"
+	ModeCosignKey                = "cosign-key"
+	ModeCosignAttestationKeyless = "cosign-attestation-keyless"
+	ModeCosignAttestationKey     = "cosign-attestation-key"
 
 	ModePolicyAny = "any"
 	ModePolicyAll = "all"